processing/requirements.txt
apache-flink==1.18.0
kafka-python==2.0.2
clickhouse-driver==0.2.6
redis==5.0.1
